# example/docs/swagger.yaml
basePath: /api/v1
definitions:
  main.LoginRequest:
    properties:
      password:
        example: "123456"
        type: string
      username:
        example: admin
        type: string
    type: object
  main.LoginResponse:
    properties:
      expires_in:
        example: 7200
        type: integer
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  main.Response:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: success
        type: string
    type: object
  main.UploadResponse:
    properties:
      filename:
        example: avatar.png
        type: string
      size:
        example: 102400
        type: integer
      url:
        example: https://example.com/uploads/avatar.png
        type: string
    type: object
  main.User:
    properties:
      age:
        example: 25
        type: integer
      email:
        example: zhangsan@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: 张三
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server.
  title: Example API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录获取 Token
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.Response'
            - properties:
                data:
                  $ref: '#/definitions/main.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response'
      summary: 用户登录
      tags:
      - Admin-Auth
  /auth/logout:
    post:
      description: 用户登出，使 Token 失效
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response'
      summary: 用户登出
      tags:
      - Admin-Auth
  /upload/avatar:
    post:
      consumes:
      - multipart/form-data
      description: 上传用户头像图片
      parameters:
      - description: 头像文件
        in: formData
        name: file
        required: true
        type: file
      - description: 用户ID
        in: formData
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.Response'
            - properties:
                data:
                  $ref: '#/definitions/main.UploadResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response'
      summary: 上传头像
      tags:
      - Upload
  /upload/files:
    post:
      consumes:
      - multipart/form-data
      description: 批量上传多个文件
      parameters:
      - description: 文件列表
        in: formData
        name: files
        required: true
        type: file
      - description: 文件描述
        in: formData
        name: description
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/main.UploadResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response'
      summary: 批量上传文件
      tags:
      - Upload
  /users:
    get:
      consumes:
      - application/json
      description: 获取所有用户的列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/main.User'
                  type: array
              type: object
      summary: 获取用户列表
      tags:
      - Admin-User
    post:
      consumes:
      - application/json
      description: 创建一个新用户
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.Response'
            - properties:
                data:
                  $ref: '#/definitions/main.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response'
      summary: 创建用户
      tags:
      - Admin-User
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定用户
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.Response'
      summary: 删除用户
      tags:
      - Admin-User
    get:
      consumes:
      - application/json
      description: 根据ID获取用户详情
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.Response'
            - properties:
                data:
                  $ref: '#/definitions/main.User'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.Response'
      summary: 获取单个用户
      tags:
      - Admin-User
    put:
      consumes:
      - application/json
      description: 更新用户信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.Response'
            - properties:
                data:
                  $ref: '#/definitions/main.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response'
      summary: 更新用户
      tags:
      - Admin-User
swagger: "2.0"